# server/swagger-auto-generated.yaml
openapi: 3.0.2
info:
  title: Campus Events and Entertainment Center API
  version: 1.0.0
  description: API for managing campus events and entertainment
servers:
  - url: http://localhost:5000/api
  - url: http://217.142.229.202/api
  - url: http://10.27.41.93:5000/api
security:
  - bearerAuth: []
paths:
  /comments/{event_id}:
    post:
      tags:
        - Comments
      summary: Create a comment
      parameters:
        - $ref: '#/components/parameters/path_event_id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
                - content
                - rating
              properties:
                content:
                  type: string
                  description: The content of the comment
                  minLength: 1
                  maxLength: 500
                rating:
                  type: integer
                  description: The rating of the event
                  minimum: 1
                  maximum: 10
      responses:
        '201':
          description: Comment created successfully
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '500':
          $ref: '#/components/responses/500'
    get:
      tags:
        - Comments
      summary: Get all comments of an event
      parameters:
        - $ref: '#/components/parameters/path_event_id'
        - $ref: '#/components/parameters/query_limit'
        - $ref: '#/components/parameters/query_offset'
      responses:
        '200':
          description: Comments found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Comment'
                properties:
                  total:
                    type: integer
                    description: The total number of comments
        '404':
          $ref: '#/components/responses/404'
          description: Event not found
        '500':
          $ref: '#/components/responses/500'
  /comments/{comment_id}:
    delete:
      '204':
        description: Comment deleted successfully
      '401':
        $ref: '#/components/responses/401'
      '403':
        $ref: '#/components/responses/403'
      '404':
        $ref: '#/components/responses/404'
      '500':
        $ref: '#/components/responses/500'
      tags:
        - Comments
      summary: >-
        Delete a comment - A user could delete only his/her own comment, while
        an admin could delete any comment
      parameters:
        - $ref: '#/components/parameters/path_comment_id'
      responses: null
...
components:
  parameters:
...
    path_event_id:
      name: event_id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
...
    path_comment_id:
      name: comment_id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    query_limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    query_offset:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Comment:
      type: object
      additionalProperties: false
      required:
        - id
        - content
        - user
        - event
        - likes
        - dislikes
        - rating
      properties:
        id:
          type: integer
          readOnly: true
          description: The auto-generated id of the comment
        content:
          type: string
          description: The content of the comment
        user_id:
          type: integer
          description: The id of the user who created the comment
        event_id:
          type: integer
        likes:
          type: integer
          default: 0
        dislikes:
          type: integer
          default: 0
        rating:
          type: integer
...
    ErrorDetail:
      title: ErrorDetail
      additionalProperties: false
      required:
        - title
        - description
      type: object
      properties:
        title:
          title: Title
          type: string
        description:
          title: Description
          type: string
...
  responses:
    '400':
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDetail'
    '401':
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDetail'
    '403':
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDetail'
    '404':
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDetail'
...
    '500':
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDetail'
...